+++ flutter/test/flutter_enricher_test.dart
@@ -216,7 +216,9 @@ void main() {
             screenHeightPixels: 1080,
             screenWidthPixels: 1920,
             screenDensity: 2,
-            theme: 'sentry_theme',
+          ),
+          operatingSystem: SentryOperatingSystem(
+            theme: 'dark',
           ),
         ),
       );
@@ -246,8 +248,8 @@ void main() {
         fakeEvent.contexts.device?.screenDensity,
       );
       expect(
-        event.contexts.device?.theme,
-        fakeEvent.contexts.device?.theme,
+        event.contexts.operatingSystem?.theme,
+        fakeEvent.contexts.operatingSystem?.theme,
       );
     });